Add missing functions, add missing controllers

+++ src/main/java/com/dropchat/cinemaxmovie/service/TicketService.java
@@ -1,15 +1,17 @@
 package com.dropchat.cinemaxmovie.service;
 
+import java.util.Random;
+
+import org.springframework.beans.BeanUtils;
+import org.springframework.stereotype.Service;
+
 import com.dropchat.cinemaxmovie.configuration.ApplicationConfig;
 import com.dropchat.cinemaxmovie.entity.Ticket;
 import com.dropchat.cinemaxmovie.repository.BillTicketRepository;
 import com.dropchat.cinemaxmovie.repository.ScheduleRepository;
 import com.dropchat.cinemaxmovie.repository.TicketRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.BeanUtils;
-import org.springframework.stereotype.Service;
 
-import java.util.Random;
+import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
@@ -29,11 +31,11 @@ public class TicketService {
     public Ticket addNew(Ticket newTicket) {
         newTicket.setCode(generateCode());
         ticketRepository.save(newTicket);
-        var current = ticketRepository.findByCode(newTicket.getCode())
+        var current = ticketRepository
+                .findByCode(newTicket.getCode())
                 .orElseThrow(() -> new RuntimeException("Date not found !"));
         scheduleRepository.findAll().forEach(schedule -> schedule.getTickets().forEach(x -> {
-            if (current.getSchedule().equals(x.getSchedule()))
-                current.setPriceTicket(schedule.getPrice());
+            if (current.getSchedule().equals(x.getSchedule())) current.setPriceTicket(schedule.getPrice());
         }));
         // Check Seat and Schedule Active before create
         ticketRepository.save(current);
@@ -41,20 +43,18 @@ public class TicketService {
     }
 
     public Ticket remake(Ticket remakeTicket) {
-        var current = ticketRepository.findById(remakeTicket.getId())
+        var current = ticketRepository
+                .findById(remakeTicket.getId())
                 .orElseThrow(() -> new RuntimeException("Date not found !"));
-        if ((int) remakeTicket.getPriceTicket() == 0)
-            remakeTicket.setPriceTicket(current.getPriceTicket());
+        if ((int) remakeTicket.getPriceTicket() == 0) remakeTicket.setPriceTicket(current.getPriceTicket());
         BeanUtils.copyProperties(remakeTicket, current, config.getNullPropertyNames(remakeTicket));
         return ticketRepository.save(current);
     }
 
     public Ticket delete(String code) {
-        var current = ticketRepository.findByCode(code)
-                .orElseThrow(() -> new RuntimeException("Date not found !"));
+        var current = ticketRepository.findByCode(code).orElseThrow(() -> new RuntimeException("Date not found !"));
         billTicketRepository.findAll().forEach(x -> {
-            if (x.getTicket().getCode().equals(code))
-                billTicketRepository.delete(x);
+            if (x.getTicket().getCode().equals(code)) billTicketRepository.delete(x);
         });
         current.setActive(false);
         return ticketRepository.save(current);
